requirements.txt
google-api-python-client==2.17.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.5
pygithub==1.55
python-dotenv==0.19.0
requests==2.26.0

